feat : 아이템 적용 및 추가 작업

+++ Assets/ScriptableObjects/Scripts/PaddleSO.cs
@@ -4,6 +4,6 @@
 public class PaddleSO : ScriptableObject
 {
     [Header("Paddle Info")]
-    public int size;
+    public float size;
     public float score;    
 }
